added breakUpText method that generates break up mssg
